refactor aziotd config
